account for overflow into DES
